created core component
